ML/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.88.0
httptools==0.5.0
pydantic==1.10.2
python-dotenv==0.21.0
starlette==0.22.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
cachetools==4.2.4
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.6
comtypes==1.1.14
datasets==2.8.0
dill==0.3.6
filelock==3.8.2
frozenlist==1.3.3
fsspec==2022.11.0
grpcio==1.51.1
grpcio-status==1.48.2
h11==0.9.0
h2==3.2.0
happytransformer==2.4.1
hpack==3.0.0
hstspreload==2022.12.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.11.1
hyperframe==5.2.0
idna==2.10
multidict==6.0.4
multiprocess==0.70.14
numpy==1.24.0
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
protobuf==3.20.3
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pypiwin32==223
pytesseract==0.3.10
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2022.7
pywin32==305
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
responses==0.18.0
rfc3986==1.5.0
rsa==4.9
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tokenizers==0.13.2
torch
tqdm==4.64.1
transformers==4.25.1
typing_extensions==4.4.0
urllib3==1.26.13
xxhash==3.1.0
yarl==1.8.2
